move collection keys methods to seperate module to not create dependecy cycle
