+++ lib/OnyxStorageManager/index.ts
@@ -2,6 +2,7 @@ import bindAll from 'lodash/bindAll';
 import type {OnyxKey} from '../types';
 import Storage from '../storage';
 import * as Logger from '../Logger';
+import {isCollectionKey, isKeyMatch} from '../CollectionKeyUtils';
 import type {StorageUsageConfig, StorageKeyInfo, StorageMetadata, StorageCleanupResult, CleanupExecutionResult} from './types';
 import {DEFAULT_STORAGE_CONFIG} from './types';
 
@@ -215,10 +216,7 @@ class OnyxStorageManager {
 
     private isKeyEvictable(key: OnyxKey): boolean {
         return this.evictableKeys.some((pattern) => {
-            if (pattern.endsWith('_')) {
-                return key.startsWith(pattern);
-            }
-            return key === pattern;
+            return isKeyMatch(pattern, key);
         });
     }
